requirements.txt
Django==1.8
gunicorn==19.3.0
djangorestframework
tweepy